requirements.txt
msal==1.31.1
requests==2.32.3
python-docx==1.1.2
openai>=1.55.0
httpx>=0.27.0
python-dotenv==1.0.1
